requirements.txt
openai~=2.20.0
python-dotenv~=1.2.1
google-search-results~=2.4.2
requests~=2.28.0